Allow user to process a given time range of packets
I'm abusing this to "parallel" process a hour-long pcap file. Each corsaro is given a different time window. The downside is that each process has to spend time seeking through the packets up to the time window, but the upside is that each process should be much faster than processing the entire hour at once. Hopefully.