Remove 'class' column from Avro schema
This change discards some information from the original FlowTuple data, but it is information that we have never used, and do not plan to capture in the new STARDUST architecture. This does however mean that tuples are now no longer unique within a 1-minute bin. We could consider using Spark to unique and re-write the Avro files if this becomes a problem (or we find we can save disk space).