Remove old corsaro from build system
The addition of the wdcap plugin has created a divergance between
old and new corsaro that will require effort to ensure old corsaro
still builds.

Rather than try to maintain both versions in the same repo, we
should fork the two versions into their own repos.

If that is going ahead, then there's no need for the libcorsaro3
branch to continue to build the old code.