First phase of implementing the DOS plugin
Only the processing API is actually present at the moment
(and is untested!) -- the merging API is just an empty shell.

There's also a lot of code left lying around from a previous
attempt that was based on the old "interim-file" style of merging.
This will get tidied up once the merging API is complete, as a
lot of this code will probably become part of the output phase of
the merging.