Ensure PPM bucket lists are populated with trailing zeroes.
In normal operation, the PPM buckets for a vector are only
updated if a packet is seen that matches the vector. However, PPM
merging works best if all bucket lists contain the full set of
buckets for the interval (even if those buckets contain zeroes).

This change will use the interval end time to push any needed
extra '0' buckets on the end of the PPM list for a vector just
prior to the vector being sent to the merging thread.