Added ability to store packet timestamps for each vector.
These can be used to calculate inter-arrival times once all
thread outputs for that vector have been merged.

At the moment, the IATs are just dumped to standard out using
a space-separated format, pending a decision on how we are
going to use the set of IATs to derive a useful method of
aggregation for the IAT measurements.

Note, this will generate a lot of standard output so you'll want
to comment out the printf in corsaro_dos.c if you're trying to
do anything other than collect data for the IAT aggregation
analysis!