Add filteringstats plugin
The plugin tracks the effectiveness of each individual subfilter
used by libcorsaro. Every interval, the plugin reports the number
of packets and bytes that would have been removed by each filter.
It also reports the number of unique source and destination IPs
appearing in the packets that would have been filtered.

There is also scope for using the plugin to evaluate other
user-provided filters. The code for this is not implemented yet,
but the appropriate places to insert the code are marked in the
plugin source with TODOs.