Added memory handler API to libcorsaro3
The memory handler is designed to optimise scenarios where we
are frequently allocating and freeing small fixed-size amounts
of memory (e.g. items for populating hash tables / sets).

If there are a lot of items or a lot of membership churn, our
program will be spending a lot of time doing repetitive calls to
malloc and free, which can have significant overhead.

The aim of this API to allocate these items in larger batches,
such that instead of doing 1 malloc 1000 times, we always alloc
1000 items at a time so we only need a single malloc.

The other side of the handler is that batches of items are
returned to a freelist once all of the items have been released
back to the handler. We can then reuse that batch instead of doing
a malloc next time we need a new batch.

Should be very helpful for plugins that have to keep track of
a lot of state, especially the report plugin.