report plugin: spread merge workload over multiple threads
With telescope data, the volumes of IP addresses that are observed
results in a massive workload for the corsaro merge thread; too
much work for one thread to be able to complete in a reasonable
amount of time.

However, the IP tallying work is easy to parallelise -- we can
group each IP address into its parent /16 and have a series of
worker threads that do the tallying for each /16 in turn. The
number of addresses matching each metric in each /16 can then be
combined to get the correct total across the whole source or
destination address set.

Next job: test this theory on the real telescope data...