Attempted performance improvements for report plugin
Try to reduce the number of times we attempt hash table lookups.

Removed 'activemetrics' map altogether, because we can get by
without it.

Combined srcips and destips IP block sets into a single "knownips"
set. Each metric for an IP now has a bit which is set if the IP
is seen as a source or dest IP for that metric.

This should also reduce the number of merging "jobs" since source
and dest IP counting can be done via the same job.