Tweak blob sizes for IP memhandlers.
Larger blob sizes should mean less mmap/munmap calls in the
memory handler, which is good in terms of performance.

Downside is that we might use a bit more memory than we really
need, especially in cases where we only end up using a handful
of entries from the most recent blob. Also, more likely to
run into mmap issues if the memory is overly fragmented.