Improve report plugin performance by batching messages
The locking required for each message was costing us tons
of time, so now I send one message with 50 results in it
instead of 50 messages with 1 result each.

Batch size could become configurable if need be.

Note: batches are currently malloced and freed, but could
easily use a memory handler to further boost performance. Will
try to look into this once I've done some validation of the
results.