Update tracker results once 2/3rds of proc threads have reported
Previously we did this as soon as one proc thread reported, but
this might be a little trigger-happy. Last commit accidentally
left in a change where I was testing what happens if we wait for
all threads to report -- probably not sustainable without ticks
to wake up inactive threads, so I've set it to 2/3 for now.