Further optimisation of report plugin.
Don't use a hash map for per-IP metric tracking until we've
seen at least 20 unique metrics for that IP. Up until then,
just use a simple static array. This saves us a lot of hash
map overhead for the majority of IPs which only have a handful
of metrics.

Also, don't look up the IP in the IP hash for each tagged metric.
The IP is not changing, so just look it up the first time around
and reuse the reference for all remaining metrics.

Removed some of the temporary debug output from previous commits.