Remove met_counters from report plugin.
Instead, we now keep track of packets and bytes per metric in
the corsaro_metric_ip_hash_t structures that are maintained by the
iptracker threads.

This saves us some hash table maintenance and lookups, although it
appears to only have a small effect (if any) on processing time
since the removed hash maps are in less busy threads.

Also added a constant for the number of metrics that can be
associated with an IP before we create a khash map.