Use khash for metric tally hashmap in report plugin.
khash seems to be a bit faster for dealing with this data.

NOTE: I also tried using khash for the known IP hash maps, but
performance was *much* worse so UThash is still the best option
here.

Also, removed add_new_message_tag() function and just put the
code directly in the function that used to call it; might save
some function overhead...