Add "nolock" versions of memhandler get and release functions.
For use cases where the memory items are never going to leave the
scope of a single thread, we don't need the overhead of locking
and unlocking a mutex every time someone wants a new memory item.

Therefore, we can add a fast version of the memhandler API for
these cases that doesn't bother with the mutex.