Add a filtering API function that will apply multiple filters.
The idea is that this function can pre-process the packet once
(to get IP headers, TCP headers, port numbers etc), then use the
internal versions of the apply functions for each filter that
the caller cares about.

In theory, this should let us get the results for multiple filters
faster -- useful for tagging, where we apply 3 high level filters,
and the filter stats plugin, which applies all of the filters.

Updated the tagger to make use of the new API function.