Set ZMQ_IMMEDIATE option to avoid queuing messages for inactive clients.
If we don't do this, a connecting client can end up receiving a
batch of messages dating back to the last time someone was
subscribed to the zmq socket. If this was long enough in the past,
there'll then be a big gap in messages (where the rest of the
backlog was dropped) followed by new recent messages.

This is a bit of a nuisance from an analysis perspective, since
we can end up with a bunch of "empty" intervals preceded by a
partial interval. Better for all of those messages to have been
dropped and simply start afresh from the time that the client
connects.