Make sure we always write a value for the ipmeta flowtuple fields.
Even if the corresponding provider was not active on the tagger,
we need to put in some sort of "unknown" value otherwise parsers
will not be able to read our avro files.

We *could* instead re-write our schema to allow these fields to
be null, but that is actually more annoying to encode than just
putting in a static '??'. It's probably also easier to recognise
on the decoder side too.