Replaced protobuf serialisation with just sending raw structures
In the current model, tagged packets shouldn't be sent outside the
system that they were created on so we don't actually need the
overhead of a serialisation format (a la protobufs). Instead,
we can just blast the original structures (packed, of course)
onto our queue and trust that the other end will be able to easily
interpret the received data as such.

Once we do decide to export tagged packets to external clients,
we should probably do so via a proxy that serialises the packets
into a standardised format -- this means that serialisation
overhead doesn't get in the way of tagged packets reaching our
own internal analysis programs running on the same host. That
serialisation proxy may want to re-use the old protobuf schema,
so I'll leave it around for now.

I have removed the protobuf autoconf checks and automake linking
for now, just to keep things somewhat tidy in term of the current
build.