Move filter tag bits into the corsaro tagged packet header.
Previously we were sending them separately as the first part of
a multi-part message. However, the cost of doing a zmq_send() for
two bytes seems to be the same as the cost of doing a zmq_send()
for a whole packet, so we're better off combining the parts into
a single message and calling zmq_send() once.

As long as the filter bits are at the front of our structure,
the subscription model for zmq works the same so everything
appears to work out.