Add separate corsarowdcap tool
Corsarowdcap reads packets via a libtrace input (just like
corsarotagger) but produces output based on intervals and
file rotations (just like a corsarotrace plugin). Since the
two methods are not entirely compatible, it was easier to split
it out into its own program.

This also means that, in theory, the wdcap capture is not dependent
on tagging etc. being able to keep up with the incoming packets.