Attempt to use async IO for fast trace writing
In theory, this will help us deal with incoming packets faster
in corsarowdcap and hopefully reduce the likelihood of packet
drops.

Also added a couple of functions that attempt to change the I/O
priority for a thread -- we don't make use of them right now but
they're there if we do decide they are useful.