Add corsaro_tag_ippayload API function
This function is the equivalent of corsaro_tag_packet(), except
that it doesn't require a libtrace packet -- it works with simply
a pointer to the IPv4 header and an amount of bytes in the packet
(starting from that header).

This is useful in situations where we are working with a copy
of the packet contents, having released the original captured
libtrace packet.

Added a new filterbit for the high level filtering: "not IP".
This will allow downstream clients to recognise non-IP packets
and discard them if necessary.