Experimental change to tagger to split tagging and capture threads
The new approach is to have a set of packet processing threads
as before, which solely deal with copying the packet into the
appropriate part of the output buffer. The tagging is now done in
a separate set of threads.

The aim is to minimise packet drops due to the tagging taking too
long to keep with the packet rate (at very high rates). If we can
do the absolute minimum work on the "libtrace" packets, we can
return the buffer space occupied by those packets to the capture
device sooner. This, of course, assumes that the tagging is the
bottleneck -- this may not actually be the case.

This change does mean that our tagger is more complex than before,
using more threads and internal queues to do the same job. All of
the extra bits may actually be detrimental to performance in the
long run, so we should not be afraid to revert these changes if
there is no tangible benefit to them.