Add consterfframing option to corsarotagger and corsarowdcap.
This is an optimisation that allows to tell libtrace that the
any ERF packets we receive are going to have a certain framing
length, so don't worry about trying to calculate it based on
the link type and ERF packet type.

It's a bit hacky but saves us a surprising amount of processing
time, which is critical for the telescope deployment.

Defaults to 16, which is a typical Ethernet frame (with no
ERF extension headers). Set it to -1 to force libtrace to
calculate the framing length.