Add optional stats file output alongside merged file
If enabled (using the new `writestats` option), a "stats" file will be written alongside the merged trace file at the end of each interval. The stats file includes per-thread and overall stats as reported by libtrace, as well as the amount of time that the merge operation took. This does require that the processing threads query libtrace for stats, and send a little more data back to the merge process, but this overhead should be minimal.