Improve internal zmq performance by moving buffer references
Previously, we were using zeromq to pass around the contents of
internal buffers which led to a lot of memcpying inside zeromq.

Now, the only thing we put in the internal queues is a pointer to
the buffer object so each buffer that is queued only results in
a copy of a pointer.

Note, I've disabled the actual final export of the tagged packet
to any interested subscribers -- this will need to be re-enabled
once I'm happy with the performance gains overall.