corsarotrace: skip writing output for incomplete interval if halted.
The data for that final interval is probably not very useful
anyway, as the interval is incomplete.

This also means that a user who is halting a corsarotrace
instance doesn't have to wait for a final merge step to complete
before the program actually halts -- depending on the plugin,
this can take anywhere up to 40 seconds and is really tedious when
doing development work especially...

This could be reverted once code is more stable and therefore less
actively developed, if required.