Ensure tagged packets emitted by corsarotagger are in chronological order
Once I added worker threads for tagging, we lost the guarantee that
the tagged packets would be in a strict chronological order.
Unfortunately, many of our corsarotrace plugins expect this to
be the case (e.g. for recognising interval boundaries) and so
these plugins were producing slightly incorrect results.

Now worker threads emit the tagged packets into their own
individual zeromq queue and the "proxy" thread for the publishing
socket now uses a priority queue to ensure that the proxy serves
the queue with the oldest packet in it at all times.

A few potential caveats to bear in mind:
 * obviously the pqueue adds extra workload -- this is in its own
   thread so hopefully won't cause any slowdowns but if servicing
   the pqueue is slow then we do have a potential bottleneck.
 * the pqueue system assumes that *all* worker threads are going
   to be busy tagging packets. If one stops or pauses for some
   reason, we could end up blocking waiting for that queue to
   produce a new packet.