Add libtimeseries configuration parsing to corsarotrace
Update report plugin to recognise that these option exist and
dump their current settings when the plugin starts up.

Fix bug with DBATS flags not actually working like a bitmask
should.

Added default channelname, compresscodec and topicprefix to
Kafka config if a broker URI is set but any of these other
options are missing.

Added "clone" API functions for libtimeseries backend option
management -- we don't actually use it in our plugins, but the
code is there in case we do decide we need a deep copy of the
options at some later point.