Add more in-code doc to libcorsaro_libtimeseries code
In the header file, I've only bothered to document the ASCII
API functions for now -- it should be easy to infer that the same
documentation applies to the corresponding functions for each of
the other backends (just with the correct name substituted for
ASCII).

Most of the code itself is really simple and doesn't need much in
the way of detailed documentation for anyone to figure out what is
going on, especially given the error messages that are logged
whenever something does behave unexpectedly.