corsarotrace: recognise intervals that are shorter than normal
A short interval can occur at either the first interval (i.e.
because the program did not start exactly on an interval
boundary) or at the end (i.e. the program was halted in the
middle of an interval).

Depending on the plugin, it may be undesirable to publish the
results for incomplete intervals as the results could be misleading
when compared against the other longer intervals.

Any intervals ended by corsarotrace are now marked with a flag
that indicates whether the interval is complete or not. Each
plugin can choose how to behave in the incomplete case -- they
can now discard their current data if it makes sense to do so.

Existing plugins have been updated to react to incomplete
intervals as follows:

 * report -- discard
 * flowtuple -- publish
 * dos -- publish
 * filteringstats - discard