Re-write corsarowdcap restarting code to better suit systemd
systemd monitoring only watches a single parent process id, so
our old method of replacing the running corsarowdcap process with
a new one whenever we are restarted didn't really work out.

Instead, corsarowdcap now always maintains a parent process that
simply waits for signals and a child process that does the capture
work. When the parent gets a signal, it forwards it on to the
child to either trigger a restart or a halt. In the case of a
restart, the parent also forks itself to create a fresh capture
process.

By this method, we maintain a constant parent process for systemd
to monitor, while still being able to temporarily have two active
wdcap processes (one merging old output, the other capturing the
latest packets) whenever we need to restart due to a config change
or similar.