corsarotagger: SIGHUP will cause the IP meta source files to be reloaded
This allows us to tell a running process to update its internal
IPmeta data structure to match the current data files on disk,
so that changes in the source meta data can be applied to the
tagger without having to restart it.

Assumes that the names of the IP meta files are the same as
they were when the tagger started -- ideally, the tagger would
be configured to load IP meta from symlinks with the name
"maxmind-latest.csv" and the symlink would be updated to point
to any new data files before triggering the HUP.

Reloading IP meta data happens in a separate thread, so packet
processing and tagging is (mostly) unaffected. The tagging will
continue to use the old data right up until the loading is
complete. Once all threads have updated to the new meta data,
the old one will be destroyed.

Other pertinent changes:
 * Wrapped tagged packet buffer messages with a message type, so
   we can send other types of messages from packet processing
   thread to tagger workers -- turned out to be unnecessary for
   this application but may be useful down the road.
 * Removed redundant corsaro_enable_ipmeta_provider() function.
 * Fix race condition where corsarotagger could hang on exit inside
   a blocking send.