trace output: replace obscure aio_return assertion failure
When the disk is full when writing traces, we used to assert fail
because an asynchronous write did not manage to write the full
expected amount but there was no real indication that the full
disk was the problem.

Now, a partial write won't cause an assertion failure -- rather,
we attempt to write the remaining content again and (hopefully)
will get a proper error that will be logged with a useful strerror
explanation.

This has the added benefit of making our code more robust in the
case where an async write does only manage to perform a partial
write for some other reason.