corsarowdcap: add option to have the merge thread write compressed output
Added two config options to control compression method and level.
If either is not set, then no compression will be performed.

Note that most compression methods will struggle to keep up with
high packet volumes (lz4 is a notable exception, as is gzip IF you
happen to have QAT available).