corsarowdcap: add support for multiple merging threads
The merge job for each interval is mapped to a specific merging
thread. This allows the merging workload to be spread across
multiple threads, allowing the user to spend more CPU cores
to compensate for merging processes that take longer than a
single interval to complete (e.g. when compressing the merged
output).

Moved the merge-related code into a separate source file, as
part of an attempt to tidy some of this code up a bit more.