report plugin: remove "ended report interval before..." message
This also removes the non-firing assertion.

I've replaced this with a loop that will wait until the
previous interval's results have been merged. This will at
least allow us to recover from a temporary blip, but if we are
consistently slow then we're going to end up with some backlog
problems.