corsarotagger: announce geolocation labels via the control socket.
Geolocation units (countries, regions, polygons) are identified
internally using 32 bit integers, but when we want to write
useful results to a database, we need to know what location those
numbers actually correspond to.

This information is available via libipmeta, but only the tagger
has access to that. We *need* the information, however, in our
corsarotrace plugins (especially the report plugin).

This commit upgrades the existing control socket on the tagger
to allow clients to query the tagger for a list of labels for each
geolocation identifier. The client may request the entire list or
just the most recent additions (i.e. in the case that the IPmeta
data has been updated / reloaded).

Updated the report plugin to use the control socket to maintain
a set of FQIDs for each known country and use them to properly
label country-level statistics.

Some elements are not very well tested, especially with regard to
whether things work properly in the event of an IPmeta reload,
but the basic use case seems to work well enough.

Future work will extend the label set to include regions and
polygons as well, which is a key requirement for implementing
full netacq geolocation support.

Removed temporary hard-coded country->continent mapping in report
plugin.