report plugin: try to send less internal messages during high packet rate events
Attempt to aggregate metric counts for IP addresses that appear
frequently, rather than sending one message per packet to the
IP tracker threads.

This may help with our memory blowouts when under heavy packet
load (but little increase in unique src or dest IPs).

Somewhat working right now, but the numbers are not entirely
consistent with the previous version of this code, so still some
kinks to probably work out.