report plugin: only delete IP entries once they are inactive
In theory, this could save us a bit of time if the same IPs
are showing up very frequently, rather than freeing and
reallocating the structure after every message we send to the
IP trackers.

Also, don't error alarmingly if the tagger is not responding to
label update requests -- just carry on with the possibly out of
date labels (and warn the user of course).

Increase bodycount to a uint32_t to avoid possible overflow
bugs if we end up increasing some other internal limits later
on.