corsarotrace: if the control socket fails, try to reconnect
This should help resolve endless spamming of messages about
being unable to send IPmeta update messages in situations where
the tagger has been restarted.

The problem was that we assumed our control socket would always
be valid, but actually it could fail and we'd need to try and
re-establish that connection somehow rather than just trying and
failing to use the original socket constantly.