Add configuration options for various zeromq HWMs
This will allow users to tweak how much memory their tagger
and corsarotrace instances are likely to use when under high
load -- more memory = less likelihood of drops, but of course
you don't want to run out of memory.

Unfortunately, HWMs are a very approximate tool. It's
impossible to really know how much memory a given HWM value
will use so trial and error will be necessary to find the
right values for a given setup.