Split report plugin code into separate source files
We had three logically distinct components to the plugin:
the processing threads, the IP tracker threads and the
merging thread, so splitting the code into separate source
files makes sense.

This code will now be a lot easier to maintain and understand,
hopefully.

In the process, I've also reverted most of the changes to how
the processing threads send updates to the IP tracker threads.
I've removed the IP metric aggregation (the CPU load introduced
by maintaining IP maps within the processing threads was too
high and causing a lot of flow-on issues) and gone back to
simple appending of IP + tags to each update message.

The update message construction is slightly improved compared
with the old "old" version -- mainly just in terms of writing
cleaner code and using zeromq multi-part messages to send the
message header separate to the body (so the receiver can
easily allocate the right amount of memory to fit the message
body).