Replace unique IPs per metric tracking approach
Instead of maintaining a map of IPs -> a set of metrics,
we now use a map of metric IDs -> bitmap set of observed IPs.
In theory, this should be slightly less memory hungry than our
old approach but does tend towards worse performance if the
IP bitmaps become large.

Adding more IP tracker threads can help spread this workload, but
sometimes the volume of traffic is just too high to keep up with.