report: attempt to fix occasional segfault when updating IP info header
If the msgbuffer gets re-alloced, our pointer to the IP header
would become stale and point to unusable memory. This can
cause problems since we update the number of tags field in that
header for every tag we process.

Also introduced a macro for the "process_single_tag" invocation,
as the return value checking and number of args was getting
annoying to work with for so many metrics.