Add new utility: corsarofanner
The fanner is designed to receive tagged packets over a TCP
zeromq socket and then re-publish them to an IPC socket for
general consumption by multiple corsarotrace processes on the
same host. The goal is to reduce TCP traffic between the host
running the tagger and a host running multiple analysis instances,
as previously each analysis instance was consuming network
bandwidth to suck in the exact same data as its fellow instances.

The fanner is still a bit of a WIP.