corsarotrace: detect whether input is coming from a fanner or tagger
We need to do slightly different things when we connect to a fanner
input source (one unique queue name per thread) vs a tagger
(each thread connects to the same "name").

The detection is simply based on whether the input queue is an
"ipc" type or a "tcp" type, which will work in most practical use
cases. We could consider adding an extra config option to override
the auto-detection if this ends up being a problem.