Flowtuple merging tasks can now be spread across multiple threads
This means that if we have intervals that take a long time to
merge ( > the length of the interval), other threads can pick up
the subsequent merging jobs and run them concurrently with the
long one.

This won't help if we're continually running behind, but it does
buy more time for the level of work to drop back down to a
manageable level.