flowtuple: add a few extra fields to each emitted flowtuple record
The fields that I've added are:
 * TCP SYN length -- the size of the TCP header on the initial SYN
 * TCP SYN window size -- the announced receive window size of the
                          initial SYN
 * "Is spoofed" -- a flag that is set to true if the first packet
                   matches any of the spoofing filters
 * "Is Masscan" -- a flag that is set to true if the first packet
                   matches all of the criteria for detecting TCP
                   Masscan SYNs

Updated the filtering code to add a new filter for TCP SYNs with
an announced window size of 1024, as well as extending the "large
scale scan" filter to match the combined criteria that we have
discovered for masscan so far.

Also fixed a bug in the flowtuple plugin which was causing the TCP
flags for a TCP packet to be derived incorrectly.