Corsaro avro writers now take a parameter to determine if snappy should be used.
Previously, we always tried snappy first and fell back to deflate
if it wasn't available.

The parameter has been added to corsaro_start_avro_writer() --
any external code that is using this API will need to be
updated to add an extra parameter when this function is called
(sorry!).

However, there are situations where deflate is a better option
than snappy -- file size is much smaller, for instance -- so
always preferring snappy is not necessarily desirable.

Updated flowtuple plugin to support a config option to state
whether snappy should be used (since FT is our biggest concern
space-wise). This option defaults to "no".

All other plugins that write avro are simply hard-coded to use
deflate only. We could add options for these too if necessary,
but for the most part it probably doesn't matter.