flowtuple: try to maintain sorted set of FTs per thread
This will mean that we don't need to perform a "sort job" before
trying to merge each thread's results. It doesn't seem to help
a lot in terms of time to complete the merge so far, but at least
feels like one less way in which we can get bottlenecked.