flowtuple: do not merge results, instead write interim files
Merging the flowtuples reported by each processing thread is too
much work to try and do within the plugin itself. This leads to
queue backlogs and eventual memory exhaustion.

Instead, we now simply have a additional set of "output" threads,
which simply write the results reported by each processing thread
into an interim avro file. These files can be merged later on,
out-of-band, by a separate process before uploading the final
merged result into swift.

It's not an ideal processing flow, but it's the only way I've
been able to cope with the volume of flowtuple data that we
are generating from the telescope.