Add non-static methods for encoding and decoding avro for flowtuples
These methods can be used outside of the flowtuple plugin to
work with flowtuple avro files (such as when we need to combine
two flowtuple records during a run of corsaroftmerge). Code
can include the flowtuple plugin header and link against libcorsaro
to access these methods.

Note that the flowtuple plugin header is not installed when
`make install` is invoked, so code outside of the corsaro-caida
ecosystem will find it harder to use these methods right now.

Also moved the flowtuple avro schema into the header file. This
makes it easier to create an avro writer that uses the same schema.