Add "nullplugin" plugin
The plugin does nothing except count packets received by each
of its processing threads. It's useful for profiling performance
issues because it forces packets to go through
the tagger's queues and into corsarotrace (as the plugin is a
client that wants them) while being able to consume the packets as
fast as our code allows since it does not do any processing or
merging work.