filtering: add corsaro_apply_all_filters() function
This function can be (in theory) a bit more efficient than
corsaro_apply_multiple_filters(), since it doesn't need to worry
if a particular filter type is *not* desired. It can also make
better use of the specific filter results when determining the
result of a broader filter category, rather than running the
same filter again.