Fix "Socket operation on non-socket" errors in report plugin
Setting standard plugin options now needs to happen *after*
corsarotrace is told by the tagger the number of processing
threads it needs to run.

Otherwise, we get situations where the report plugin doesn't
initialise enough queues for the processing threads to push
results to the IP tracker threads.

Also added an extra log message to indicate when corsarotrace
is trying to talk to the tagger over the control socket -- if
the control socket is mis-specified in the config file,
corsarotrace will hang here and without this message it is not
obvious why it is hanging.