Add faux "tagger control" socket to corsarotrace
This will be used to perform the ipmeta label lookups that would
normally be done as queries to a corsarotagger instance via the
control socket.

Added a "halt faux" message to the tagger control socket protocol,
which is used by corsarotrace to halt the thread that is running
the faux tagger control socket. Updated the corsarotagger version
of the socket to ignore these messages.