corsarotrace: add option to write packets received, dropped, etc. to a file
The "stats" file works the same way it does for corsarotagger.

For each interval, each thread writes its stats for that interval
to a uniquely named stats file. The stats are overwritten at each
interval, so if they aren't scraped before the next interval then
they will be lost (this copies the tagger behaviour).

The stats that corsarotrace will write are:
 * interval timestamp
 * number of packets received by that thread
 * number of packets detected as dropped by that thread
 * number of instances of packet loss detected by that thread
   (a loss instance may result in multiple packets being lost)
 * number of packets seen that should have been seen during an
   earlier inteval (indicates packet reordering is happening).