report: fix bug where unique source ASNs could be under-counted
If a processing thread reported 0 ASNs for a metric, then the
ASN count would be set to zero (even if the count had already
been incremented by the results from other processing threads.

This was most obvious when looking at source ASN counts for the
per-ASN metrics -- these should always be 1, but were often
being reported as 0.