report: use simpler data structures for tracking certain metrics
The aim here is to reduce the map lookup + insertion workload for
metrics where a map is possibly overkill.

The metrics updated are:
 * combined -- looking up the same id in a map continuously is
               pointless, just track combined as a single
               standalone entry.
 * IP protocols -- only 256 possible values, just create an array
                   instead.
 * Filtering criteria -- metric range is limited and consecutive,
                         so use an array instead.
 * TCP ports -- almost all ports are used, so use an array for
                storing these metrics.

Other metrics are either sparse (e.g. ICMP type-code pairs), or
have a large range that doesn't suit an array implementation
(e.g. ASNs). Only a portion of UDP ports are used in any given
interval, so I've left that as part of the "general" map.

I have other plans coming up for dealing with lookups for
geolocation metrics -- watch this space.