report: reduce metric map lookups for netacq geolocation tags
This change takes advantage of the idea that netacq tags are
hierarchical, i.e. a county will always belong to the same region,
country and continent.

This means that in our metric map, we do not need to store
entries for all four tags assigned to a packet -- we just store
the tag that is furtherest down the hierarchy, but include its
parent tags as part of the map entry. So instead of doing
up to 4 lookups per packet, we just do one.

When it comes time to merge the results, then we can then use the
stored parent tags to make sure we aggregate together the results
for all the child counties to produce the correct result for a
region, all the child regions to produce the country results, etc.

End result: our IP tracker threads should use a lot less CPU when
tracking netacq location stats, but still produce the same results
as before.