Fix issue where the last report interval was always considered incomplete
When replaying from a pcap trace, the last interval is actually
often complete but the packets that would normally trigger that
interval to end would have been written to the "next" pcap file.

Since the report plugin only writes complete intervals to its
output, we need to catch a case where the interval duration is
all but complete when we reach the end of a trace file input.