corsarotrace: warn rather than fail if "outtemplate" config is missing
Some use cases will not write output files (e.g. the report plugin
when writing directly to libtimeseries), so we shouldn't force
users to come up with a template that is never going to be used.

However, we should at least try to warn them in case they are
using a plugin / output format that will write to avro and
forget to provide a template -- should make the cause of the
errors that they are going to get much more obvious.