report: add ability to limit port series to specific ranges
Tracking stats for all 65536 TCP and UDP ports (both source
and dest) will generate over 250,000 different datapoints
per interval. This places a huge burden on our data storage
backend, so it is useful to have the option to limit the results
to just particular ports that we care about.

Port ranges are configured by adding a list of ranges to
the plugin config under the appropriate option name, e.g.

      tcp_dest_port_range:
        - 0 - 1024
      tcp_source_port_range:
        - 0 - 2048
        - 10000 - 13000
      udp_source_port_range:
        - 53 - 53
      udp_dest_port_range:
        - 0 - 10000

A specified port range is treated as inclusive; a range of
0 - 1024 will include both port 0 and port 1024, as well as
all ports in between.

If a port range option is not specified, all ports will be
included in the result set for that metric -- e.g. no
`tcp_dest_port_range` option will generate results for all
TCP destination ports.